Create all the managment slaves.
add 3 classes:
1. AbstractSlave (hold useful methods that being used by almost all slaves)

2. ALL_ID_Manager - explanation inside.

3. EnumStringConventor - to convert between enum and string (I'm trying to kill moduleGlobalHandler)
